feat(auth): margin pre-admit on admin passkey and OAuth free intake
GAP-256 PR-3b: share admitFreeIntake + ensureFreeSignupEntitlement via
@revealui/auth/server; wire admin password sign-up, passkey register-verify,
and OAuth callback for new users (K13 pre-admit + K15 free@t0).
